requirements.txt
gunicorn
django-heroku
Django
dj-database-url
sqlparse
django-bootstrap-v5
django-allauth
psycopg
psycopg2
boto3
django-storages
Pillow